disable oldstyle tests

+++ Assets/Scripts/Game/Logic/Models/Services/Constructions/BuildingService.cs
@@ -1,15 +1,10 @@
-﻿using Game.Assets.Scripts.Game.Logic.Common.Core;
-using Game.Assets.Scripts.Game.Logic.Models.Constructions;
+﻿using Game.Assets.Scripts.Game.Logic.Models.Constructions;
 using Game.Assets.Scripts.Game.Logic.Models.Entities.Constructions;
 using Game.Assets.Scripts.Game.Logic.Models.Events.Constructions;
 using Game.Assets.Scripts.Game.Logic.Models.Repositories;
 using Game.Assets.Scripts.Game.Logic.Models.Services.Resources.Points;
 using Game.Assets.Scripts.Game.Logic.Models.ValueObjects.Constructions;
-using Game.Assets.Scripts.Game.Logic.Models.ValueObjects.Resources;
-using Game.Assets.Scripts.Game.Logic.Presenters.Services.Constructions;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 
 namespace Game.Assets.Scripts.Game.Logic.Models.Services.Constructions
 {
@@ -36,13 +31,16 @@ namespace Game.Assets.Scripts.Game.Logic.Models.Services.Constructions
             if (!_constructionsService.CanPlace(card, position, rotation))
                 throw new Exception("Can't build construction here");
 
+            if (!_handService.Has(card))
+                throw new Exception("No such card in hand");
+
             var construction = new Construction(card.Scheme, position, rotation);
             var points = _constructionsService.GetPoints(card, position, rotation);
 
             _handService.Remove(card);
             _constructions.Add(construction);
 
-            _pointsService.ChangePoints(points, _fieldService.GetWorldPosition(construction));
+            _pointsService.Change(points, _fieldService.GetWorldPosition(construction));
 
             _constructions.FireEvent(construction, new ConstructionBuildedByPlayerEvent());
